RAG_Real_life_use_case_Week3_Day5/Assignment-TamannaYadav/requirements.txt
# Tesla RAG System Dependencies

# PDF Processing
pdfplumber>=0.10.0
PyPDF2>=3.0.0

# Embeddings
sentence-transformers>=2.2.0

# Vector Database
faiss-cpu>=1.7.4

# LLM API
groq>=0.4.0

# Data Processing
numpy>=1.24.0

# Web UI
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
